refactor: style and elements
