+++ kickFlip-project/src/components/cardList/cardList.tsx
@@ -1,5 +1,3 @@
-import { SyntheticEvent, useState } from 'react';
-
 import { ProductProjected, TransformParams } from '@/types/types';
 import Card from '../card/card';
 import './cardList.css';
@@ -11,18 +9,9 @@ interface CardListProps {
 }
 
 function CardList({ products, setCategories, categories }: CardListProps): JSX.Element {
-    const [searchTerm, setSearchTerm] = useState('');
-
     const handleSelect = (event: React.ChangeEvent<HTMLSelectElement>) => {
         setCategories((prevCategories) => {
             return { ...prevCategories, sort: event.target.value };
-        });
-    };
-
-    const handleSearch = (event: SyntheticEvent<HTMLFormElement, SubmitEvent>) => {
-        event.preventDefault();
-        setCategories((prevCategories) => {
-            return { ...prevCategories, search: searchTerm.toLowerCase() };
         });
     };
 
@@ -30,25 +19,13 @@ function CardList({ products, setCategories, categories }: CardListProps): JSX.E
         <div className="card-list-container">
             <div className="card-list-info-container">
                 <h2 className="card-list-title">{`The best kicks (${products.length})`}</h2>
-                <div className="card-list-sorting-container">
-                    <form onSubmit={handleSearch}>
-                        <input
-                            type="text"
-                            className="card-list-sorting"
-                            placeholder="Search"
-                            onChange={(event) => setSearchTerm(event.target.value)}
-                            value={searchTerm}
-                        />
-                        <button type="submit" className="card-list-search-button" aria-label="Search" />
-                    </form>
-                    <select id="sort-select" className="card-list-sorting" onChange={handleSelect}>
-                        <option value="">Sort by</option>
-                        <option value="price asc">Price (Low-High)</option>
-                        <option value="price desc">Price (High-Low)</option>
-                        <option value="name.en-US desc">Name ⭡</option>
-                        <option value="name.en-US asc">Name ⭣</option>
-                    </select>
-                </div>
+                <select id="sort-select" className="card-list-sorting" onChange={handleSelect}>
+                    <option value="">Sort by</option>
+                    <option value="price asc">Price (Low-High)</option>
+                    <option value="price desc">Price (High-Low)</option>
+                    <option value="name.en-US desc">Name ⭡</option>
+                    <option value="name.en-US asc">Name ⭣</option>
+                </select>
             </div>
             {products.length === 0 ? (
                 <p className="card-list-title">No kicks match these filters. Try modifying your search</p>
