+++ kickFlip-project/src/components/layout/basicLayout.css
@@ -6,4 +6,5 @@
     width: 100%;
     padding: 40px;
     justify-content: center;
+    max-width: 1440px;
 }
