+++ kickFlip-project/src/main.css
@@ -82,4 +82,6 @@ body {
 
 main {
     flex-grow: 1;
+    justify-content: center;
+    display: flex;
 }
