+++ kickFlip-project/src/pages/products/productsPage.css
@@ -21,7 +21,7 @@
 .products {
     display: grid;
     grid-template-columns: 1fr 3fr;
-    gap: 0 1.6rem;
+    gap: 0 4rem;
     align-self: self-start;
     position: relative;
     width: 100%;
